scripts/requirements.txt
beautifulsoup4>=4.12.0
provider-hub==0.5.0
python-dotenv>=1.0.1
